Create SPAdes_Run
This file includes the command line parameters used in the SPAdes genome assembly of sample 2-9-3. The first two files are specified as one library, but the first file is a forward-direction read of the Illumina genome, while the second file is the reverse-direction read of the Illumina genome. The third file is the Nanopore data which had already been corrected using SPAdes software.